requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
langchain==0.1.11
langchain-openai==0.0.8
chromadb==0.4.24
tiktoken==0.6.0
docx2txt==0.9
llama-index==0.7.0
llama-index-groq==0.7.0


